refactor: decompose PlayerContent.tsx into focused modules
Extract the 1,061-line monolith into five focused files under
src/components/PlayerContent/:

- index.tsx (242 lines) — thin orchestrator composing sub-components
- AlbumArtSection.tsx — album art rendering, flip state, accent color,
  bounds detection, gesture integration
- DrawerOrchestrator.tsx — queue/library drawer state, lazy imports,
  save-queue dialog, toast notifications
- GestureLayer.tsx — horizontal/vertical swipe gesture setup
- PlayerControlsSection.tsx — controls card, bottom bar, VFX menu,
  keyboard shortcuts
- styled.ts — shared styled-components

No behavior changes. All 577 tests pass.

Closes #420
