--- src/components/PlayerContent.tsx
@@ -1,1061 +1,0 @@
-import React, { Suspense, lazy, useState, useCallback, useRef, useEffect, useMemo } from 'react';
-import { clearCacheWithOptions } from '@/services/cache/libraryCache';
-import { clearAllPins } from '@/services/settings/pinnedItemsStorage';
-import type { ClearCacheOptions } from '@/components/VisualEffectsMenu';
-import styled, { useTheme } from 'styled-components';
-import { CardContent } from './styled';
-import AlbumArt from './AlbumArt';
-import SpotifyPlayerControls from './SpotifyPlayerControls';
-import BottomBar from './BottomBar';
-import { BOTTOM_BAR_HEIGHT } from './BottomBar/styled';
-import { cardBase } from '../styles/utils';
-import { ProfiledComponent } from '@/components/ProfiledComponent';
-import { usePlayerSizingContext } from '@/contexts/PlayerSizingContext';
-import { useKeyboardShortcuts } from '@/hooks/useKeyboardShortcuts';
-import { useSwipeGesture } from '@/hooks/useSwipeGesture';
-import { useVerticalSwipeGesture } from '@/hooks/useVerticalSwipeGesture';
-import { useVisualEffectsState } from '@/hooks/useVisualEffectsState';
-import { useVolume } from '@/hooks/useVolume';
-import { useLikeTrack } from '@/hooks/useLikeTrack';
-import { useTrackListContext, useCurrentTrackContext } from '@/contexts/TrackContext';
-import { useProviderContext } from '@/contexts/ProviderContext';
-import { useUnifiedLikedTracks } from '@/hooks/useUnifiedLikedTracks';
-import { LIKED_SONGS_ID } from '@/constants/playlist';
-import { useColorContext } from '@/contexts/ColorContext';
-import { useVisualEffectsContext } from '@/contexts/VisualEffectsContext';
-import { useProfilingContext } from '@/contexts/ProfilingContext';
-import { useVisualizerDebug } from '@/contexts/VisualizerDebugContext';
-import { STORAGE_KEYS } from '@/constants/storage';
-import LibraryDrawer from './LibraryDrawer';
-import AlbumArtQuickSwapBack from './AlbumArtQuickSwapBack';
-import type { AddToQueueResult, MediaTrack, ProviderId } from '@/types/domain';
-import { LIBRARY_REFRESH_EVENT } from '@/hooks/useLibrarySync';
-import Toast from './Toast';
-import RadioProgressToast from './RadioProgressToast';
-import type { RadioState } from '@/hooks/useRadio';
-import type { RadioProgress } from '@/hooks/usePlayerLogic';
-import { providerRegistry } from '@/providers/registry';
-
-const SaveQueueDialog = lazy(() => import('./SaveQueueDialog'));
-
-const QueueDrawer = lazy(() => import('./QueueDrawer'));
-const QueueBottomSheet = lazy(() => import('./QueueBottomSheet'));
-const VisualEffectsMenu = lazy(() => import('./VisualEffectsMenu/index'));
-const KeyboardShortcutsHelp = lazy(() => import('./KeyboardShortcutsHelp'));
-
-interface PlaybackHandlers {
-  onPlay: () => void;
-  onPause: () => void;
-  onNext: () => void;
-  onPrevious: () => void;
-  onTrackSelect: (index: number) => void;
-  onOpenLibraryDrawer: () => void;
-  onCloseLibraryDrawer: () => void;
-  onPlaylistSelect: (playlistId: string, playlistName: string, provider?: ProviderId) => void;
-  onAddToQueue?: (
-    playlistId: string,
-    playlistName?: string,
-    provider?: ProviderId,
-  ) => Promise<AddToQueueResult | null>;
-  onAlbumPlay: (albumId: string, albumName: string) => void;
-  onBackToLibrary: () => void;
-  onStartRadio?: () => void;
-  onRemoveFromQueue?: (index: number) => void;
-  onReorderQueue?: (fromIndex: number, toIndex: number) => void;
-}
-
-interface AlbumArtBounds {
-  left: number;
-  top: number;
-  width: number;
-  height: number;
-}
-
-interface PlayerContentProps {
-  isPlaying: boolean;
-  showLibraryDrawer: boolean;
-  onAlbumArtBoundsChange?: (bounds: AlbumArtBounds | null) => void;
-  handlers: PlaybackHandlers;
-  currentTrackProvider?: ProviderId;
-  radioState?: RadioState;
-  isRadioAvailable?: boolean;
-  radioActive?: boolean;
-  mediaTracksRef?: React.RefObject<MediaTrack[]>;
-  radioProgress?: RadioProgress | null;
-  onDismissRadioProgress?: () => void;
-}
-
-const ContentWrapper = styled.div.withConfig({
-  shouldForwardProp: (prop) => !['width', 'padding', 'useFluidSizing', 'transitionDuration', 'transitionEasing', '$zenMode'].includes(prop),
-}) <{
-  width: number;
-  padding: number;
-  useFluidSizing: boolean;
-  transitionDuration: number;
-  transitionEasing: string;
-  $zenMode?: boolean;
-}>`
-  width: ${props => props.$zenMode ? '100%' : props.useFluidSizing ? '100%' : `${props.width}px`};
-  max-width: ${props => props.$zenMode ? '100%' : `${props.width}px`};
-
-  margin: 0 auto;
-  padding: ${props => props.padding}px;
-  padding-bottom: ${props => props.$zenMode ? props.padding : `calc(${props.padding + BOTTOM_BAR_HEIGHT}px + env(safe-area-inset-bottom, 0px))`};
-  box-sizing: border-box;
-  position: relative;
-  z-index: 2;
-  overflow: visible;
-
-  transition: width ${props => props.$zenMode ? '1000ms cubic-bezier(0.4, 0, 0.2, 1) 300ms' : '1000ms cubic-bezier(0.4, 0, 0.2, 1)'},
-            padding ${props => props.transitionDuration}ms ${props => props.transitionEasing},
-            padding-bottom ${props => props.$zenMode ? '1000ms cubic-bezier(0.4, 0, 0.2, 1) 300ms' : '1000ms cubic-bezier(0.4, 0, 0.2, 1)'},
-            max-width ${props => props.$zenMode ? '1000ms cubic-bezier(0.4, 0, 0.2, 1) 300ms' : '1000ms cubic-bezier(0.4, 0, 0.2, 1)'};
-
-  container-type: inline-size;
-  container-name: player;
-  display: flex;
-  flex-direction: column;
-  /* Prevent overflow on tablets; content must fit above bottom bar + safe area */
-  max-height: ${props => props.$zenMode ? 'none' : '100dvh'};
-`;
-
-const LoadingCard = styled.div.withConfig({
-  shouldForwardProp: (prop: string) => !['backgroundImage', 'standalone', 'accentColor', 'glowEnabled', 'glowIntensity', 'glowRate'].includes(prop),
-}) <{
-  backgroundImage?: string;
-  standalone?: boolean;
-  accentColor?: string;
-  glowEnabled?: boolean;
-  glowIntensity?: number;
-  glowRate?: number;
-}>`
-  ${cardBase};
-  margin-top: 0.5rem;
-  margin-bottom: 0.25rem;
-  margin-left: 0.25rem;
-  margin-right: 0.25rem;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-  border-radius: 1.25rem;
-  border: 1px solid ${({ theme }) => theme.colors.borderSubtle};
-  box-shadow: ${({ theme }) => theme.shadows.card};
-  transition: box-shadow ${({ theme }) => theme.transitions.normal}, border-color ${({ theme }) => theme.transitions.normal};
-  ${({ theme, backgroundImage }) => backgroundImage ? `
-    &::after {
-      position: absolute;
-      inset: 0.1rem;
-      background-image: url(${backgroundImage});
-      background-size: cover;
-      background-position: center;
-      background-repeat: no-repeat;
-      border-radius: 1.25rem;
-      z-index: 0;
-    }
-    &::before {
-      position: absolute;
-      inset: 0;
-      background: ${theme.colors.card.overlay};
-      backdrop-filter: blur(40px) saturate(180%);
-      -webkit-backdrop-filter: blur(40px) saturate(180%);
-      border-radius: 1.25rem;
-      z-index: 1;
-    }
-  ` : `
-    background: linear-gradient(
-      to bottom,
-      ${theme.colors.card.gradientTop} 0%,
-      ${theme.colors.card.gradientBottom} 100%
-    );
-    backdrop-filter: blur(20px) saturate(180%);
-    -webkit-backdrop-filter: blur(20px) saturate(180%);
-  `}
-`;
-
-function ControlsLoadingFallback() {
-  const theme = useTheme();
-  return (
-    <div style={{
-      height: '80px',
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-      color: theme.colors.muted.foreground
-    }}>
-      Loading controls...
-    </div>
-  );
-}
-
-function QueueLoadingFallback() {
-  const theme = useTheme();
-  return (
-    <div style={{
-      position: 'fixed',
-      top: 0,
-      right: 0,
-      bottom: 0,
-      width: theme.drawer.widths.tablet,
-      background: theme.colors.overlay.panel,
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-      color: theme.colors.muted.foreground
-    }}>
-      Loading queue...
-    </div>
-  );
-}
-
-function VisualEffectsLoadingFallback() {
-  const theme = useTheme();
-  return (
-    <div style={{
-      position: 'fixed',
-      top: 0,
-      right: 0,
-      width: '350px',
-      height: '100vh',
-      background: theme.colors.overlay.panel,
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-      color: theme.colors.muted.foreground,
-      fontSize: theme.fontSize.sm
-    }}>
-      Loading effects...
-    </div>
-  );
-}
-
-const PlayerContainer = styled.div`
-  position: relative;
-  z-index: 1;
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-`;
-
-const PlayerStack = styled.div.withConfig({
-  shouldForwardProp: (prop) => !['$zenMode'].includes(prop),
-})<{ $zenMode?: boolean }>`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  min-height: 0; /* Allow flex shrink so content fits above bottom bar */
-  max-width: ${({ $zenMode }) => $zenMode
-    ? `min(calc(100vw - 32px), calc(100dvh - 130px))`
-    : `min(calc(100vw - 48px), calc(100dvh - var(--player-controls-height, 220px) - 120px))`
-  };
-  margin: 0 auto;
-  /* Entering zen: art grows after controls fade out (300ms delay). Exiting zen: art shrinks immediately. */
-  transition: ${({ $zenMode }) => $zenMode
-    ? 'max-width 1000ms cubic-bezier(0.4, 0, 0.2, 1) 300ms'
-    : 'max-width 1000ms cubic-bezier(0.4, 0, 0.2, 1)'
-  };
-`;
-
-const ZenControlsWrapper = styled.div.withConfig({
-  shouldForwardProp: (prop) => !['$zenMode'].includes(prop),
-})<{ $zenMode: boolean }>`
-  display: grid;
-  grid-template-rows: ${({ $zenMode }) => $zenMode ? '0fr' : '1fr'};
-  opacity: ${({ $zenMode }) => $zenMode ? 0 : 1};
-  transform: ${({ $zenMode }) => $zenMode ? 'scale(0.95) translateY(-8px)' : 'scale(1) translateY(0)'};
-  transform-origin: top center;
-  /*
-   * grid-template-rows animates over the actual element height (unlike max-height which
-   * animates over an arbitrary 500px range, causing non-linear perceived speed).
-   * Entering zen: controls collapse + fade in 300ms, then art expands (300ms delay).
-   * Exiting zen: art shrinks first (1000ms), then controls expand + fade in (350ms).
-   * --player-controls-height is pre-set synchronously via stableControlsHeightRef before
-   * the state flip, so PlayerStack has the correct target from the first animation frame.
-   */
-  transition: ${({ $zenMode }) => $zenMode
-    ? 'grid-template-rows 300ms ease, opacity 300ms ease, transform 300ms ease'
-    : 'grid-template-rows 500ms ease 500ms, opacity 1200ms ease 1200ms, transform 300ms ease 300ms'
-  };
-  pointer-events: ${({ $zenMode }) => $zenMode ? 'none' : 'auto'};
-`;
-
-// Direct grid item — min-height: 0 allows the grid row to collapse to 0fr,
-// overflow: hidden clips content during the collapse animation.
-const ZenControlsInner = styled.div`
-  min-height: 0;
-  overflow: hidden;
-`;
-
-// Zen mode track info — non-interactive song + artist label below album art
-const ZenTrackInfo = styled.div.withConfig({
-  shouldForwardProp: (prop) => !['$zenMode'].includes(prop),
-})<{ $zenMode: boolean }>`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  pointer-events: none;
-  margin-top: ${({ theme }) => theme.spacing.sm};
-  padding: 0 ${({ theme }) => theme.spacing.md};
-  overflow: hidden;
-  opacity: ${({ $zenMode }) => $zenMode ? 1 : 0};
-  max-height: ${({ $zenMode }) => $zenMode ? '5rem' : '0'};
-  transition: ${({ $zenMode }) => $zenMode
-    ? 'opacity 600ms ease 800ms, max-height 300ms ease 300ms'
-    : 'opacity 200ms ease, max-height 200ms ease'
-  };
-`;
-
-const ZenTrackName = styled.div.withConfig({
-  shouldForwardProp: (prop) => !['$isMobile', '$isTablet'].includes(prop),
-})<{ $isMobile: boolean; $isTablet: boolean }>`
-  font-weight: ${({ theme }) => theme.fontWeight.semibold};
-  font-size: ${({ $isMobile, $isTablet, theme }) => {
-    if ($isMobile) return theme.fontSize.lg;
-    if ($isTablet) return theme.fontSize.xl;
-    return theme.fontSize['2xl'];
-  }};
-  color: ${({ theme }) => theme.colors.white};
-  text-shadow: ${({ theme }) => theme.shadows.textMd};
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  width: 100%;
-`;
-
-const ZenTrackArtist = styled.div`
-  font-size: ${({ theme }) => theme.fontSize.base};
-  line-height: 1.4;
-  color: ${({ theme }) => theme.colors.gray[300]};
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  width: 100%;
-`;
-
-// Album art container with click handler
-const ClickableAlbumArtContainer = styled.div<{ $swipeEnabled?: boolean; $bothGestures?: boolean }>`
-  position: relative;
-  cursor: pointer;
-  z-index: 3;
-  perspective: 1200px;
-  filter: drop-shadow(${({ theme }) => theme.shadows.drop});
-  ${({ $swipeEnabled, $bothGestures }) => $swipeEnabled && `
-    touch-action: ${$bothGestures ? 'none' : 'pan-y'};
-    user-select: none;
-    -webkit-user-select: none;
-  `}
-`;
-
-const FlipInner = styled.div.withConfig({
-  shouldForwardProp: (prop) => !['$isFlipped'].includes(prop),
-})<{ $isFlipped: boolean }>`
-  width: 100%;
-  aspect-ratio: 1;
-  position: relative;
-  transform-style: preserve-3d;
-  transition: transform 0.6s cubic-bezier(0.4, 0, 0.2, 1);
-  transform: ${({ $isFlipped }) => $isFlipped ? 'rotateY(180deg)' : 'rotateY(0deg)'};
-`;
-
-
-const PlayerContent: React.FC<PlayerContentProps> = React.memo(({ isPlaying, showLibraryDrawer, onAlbumArtBoundsChange, handlers, currentTrackProvider, radioState, isRadioAvailable, radioActive, mediaTracksRef, radioProgress, onDismissRadioProgress }) => {
-  // --- Context hooks ---
-  const { tracks, shuffleEnabled, handleShuffleToggle, selectedPlaylistId } = useTrackListContext();
-  const { isUnifiedLikedActive } = useUnifiedLikedTracks();
-  const showProviderIcons = (isUnifiedLikedActive && selectedPlaylistId === LIKED_SONGS_ID) || !!radioActive;
-  const { currentTrack, currentTrackIndex, showQueue, setShowQueue } = useCurrentTrackContext();
-
-  const {
-    accentColor,
-    customAccentColors,
-    setAccentColor,
-    handleSetAccentColorOverride,
-    handleRemoveAccentColorOverride,
-    handleResetAccentColorOverride,
-    handleSetCustomAccentColor,
-    handleRemoveCustomAccentColor,
-  } = useColorContext();
-
-  const {
-    visualEffectsEnabled,
-    setVisualEffectsEnabled,
-    backgroundVisualizerEnabled,
-    setBackgroundVisualizerEnabled,
-    backgroundVisualizerStyle,
-    setBackgroundVisualizerStyle,
-    backgroundVisualizerIntensity,
-    setBackgroundVisualizerIntensity,
-    translucenceEnabled,
-    setTranslucenceEnabled,
-    translucenceOpacity,
-    zenModeEnabled,
-    setZenModeEnabled,
-    showVisualEffects,
-    setShowVisualEffects,
-  } = useVisualEffectsContext();
-
-  const { enabled: profilerEnabled, toggle: profilerToggle } = useProfilingContext();
-  const vizDebugCtx = useVisualizerDebug();
-  const visualizerDebugEnabled = vizDebugCtx?.isDebugMode ?? false;
-
-  // --- Leaf hooks (self-contained, fine to call here) ---
-  const { effectiveGlow, handleGlowIntensityChange, handleGlowRateChange, restoreGlowSettings } = useVisualEffectsState();
-  const { handleMuteToggle, isMuted, volume, setVolumeLevel } = useVolume(currentTrackProvider);
-  const { isLiked, isLikePending, handleLikeToggle } = useLikeTrack(currentTrack?.id, currentTrack?.provider);
-
-  // --- Save queue as playlist ---
-  const { connectedProviderIds } = useProviderContext();
-  const [showSaveQueueDialog, setShowSaveQueueDialog] = useState(false);
-  const [toast, setToast] = useState<{
-    message: string;
-    actionLabel?: string;
-    onAction?: () => void;
-  } | null>(null);
-  const saveProviders = useMemo(
-    () => connectedProviderIds.filter(id => {
-      const desc = providerRegistry.get(id);
-      return desc?.savePlaylist != null;
-    }),
-    [connectedProviderIds],
-  );
-  const canSaveQueue = saveProviders.length > 0 && tracks.length > 0;
-  const trackProviders = useMemo(() => new Set(tracks.map(t => t.provider).filter(Boolean)), [tracks]);
-  const handleOpenSaveQueue = useCallback(() => setShowSaveQueueDialog(true), []);
-  const handleCloseSaveQueue = useCallback(() => setShowSaveQueueDialog(false), []);
-
-  const handleSaveQueue = useCallback(async (name: string, provider: ProviderId): Promise<boolean> => {
-    try {
-      const descriptor = providerRegistry.get(provider);
-      if (!descriptor?.savePlaylist || !descriptor.auth.isAuthenticated()) return false;
-
-      const mediaTracks = mediaTracksRef?.current;
-      if (!mediaTracks || mediaTracks.length === 0) return false;
-
-      const result = await descriptor.savePlaylist(name, mediaTracks);
-      if (!result) return false;
-
-      setShowSaveQueueDialog(false);
-      setToast({ message: `Saved "${name}" to ${descriptor.name}` });
-      window.dispatchEvent(new CustomEvent(LIBRARY_REFRESH_EVENT, { detail: { providerId: provider } }));
-      return true;
-    } catch (err) {
-      console.error('[SaveQueue] Failed to save:', err);
-      return false;
-    }
-  }, [mediaTracksRef]);
-
-  const handleDismissToast = useCallback(() => setToast(null), []);
-
-  // --- Local UI state ---
-  const [showHelp, setShowHelp] = useState(false);
-  const [librarySearchQuery, setLibrarySearchQuery] = useState<string | undefined>(undefined);
-  const [libraryViewMode, setLibraryViewMode] = useState<'playlists' | 'albums' | undefined>(undefined);
-
-  const toggleHelp = useCallback(() => setShowHelp(prev => !prev), []);
-  const closeHelp = useCallback(() => setShowHelp(false), []);
-
-  const [isFlipped, setIsFlipped] = useState(false);
-  const toggleFlip = useCallback(() => setIsFlipped(f => !f), []);
-  const flipContainerRef = useRef<HTMLDivElement>(null);
-
-  const handleRetryAlbumArt = useCallback(async () => {
-    const providerId = currentTrackProvider ?? currentTrack?.provider;
-    if (!providerId) return;
-    const { providerRegistry } = await import('@/providers/registry');
-    const descriptor = providerRegistry.get(providerId);
-    descriptor?.playback.refreshCurrentTrackArt?.();
-  }, [currentTrackProvider, currentTrack?.provider]);
-  const albumArtContainerRef = useRef<HTMLDivElement | null>(null);
-
-  // Report album art bounds for trail visualizer (head stays inside art)
-  useEffect(() => {
-    if (!onAlbumArtBoundsChange) return;
-    const el = albumArtContainerRef.current;
-    if (!el) {
-      onAlbumArtBoundsChange(null);
-      return;
-    }
-    const report = () => {
-      const r = el.getBoundingClientRect();
-      if (r.width > 0 && r.height > 0) {
-        onAlbumArtBoundsChange({ left: r.left, top: r.top, width: r.width, height: r.height });
-      } else {
-        onAlbumArtBoundsChange(null);
-      }
-    };
-    report();
-    const ro = new ResizeObserver(report);
-    ro.observe(el);
-    return () => {
-      ro.disconnect();
-      onAlbumArtBoundsChange(null);
-    };
-  }, [onAlbumArtBoundsChange]);
-
-  // Reset flip when track changes
-  useEffect(() => {
-    setIsFlipped(false);
-  }, [currentTrack?.id]);
-
-  // Close flip when clicking outside
-  useEffect(() => {
-    if (!isFlipped) return;
-    const handleClickOutside = (e: MouseEvent) => {
-      const target = e.target as Element;
-      if (target.closest?.('[data-eyedropper-overlay]')) return;
-      const insideAlbumArt = flipContainerRef.current?.contains(target);
-      if (!insideAlbumArt) {
-        setIsFlipped(false);
-      }
-    };
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => document.removeEventListener('mousedown', handleClickOutside);
-  }, [isFlipped]);
-
-  // --- Color handlers (replaces useCustomAccentColors) ---
-  const handleCustomAccentColor = useCallback((color: string) => {
-    if (currentTrack?.albumId) {
-      if (color === '') {
-        handleRemoveAccentColorOverride(currentTrack.albumId);
-        handleRemoveCustomAccentColor(currentTrack.albumId);
-      } else {
-        handleSetAccentColorOverride(currentTrack.albumId, color);
-        handleSetCustomAccentColor(currentTrack.albumId, color);
-        setAccentColor(color);
-      }
-    }
-  }, [currentTrack?.albumId, handleSetAccentColorOverride, handleRemoveAccentColorOverride, handleSetCustomAccentColor, handleRemoveCustomAccentColor, setAccentColor]);
-
-  const handleAccentColorChange = useCallback((color: string) => {
-    if (color === 'RESET_TO_DEFAULT' && currentTrack?.albumId) {
-      handleResetAccentColorOverride(currentTrack.albumId);
-      // useAccentColor effect in usePlayerLogic will re-extract from artwork automatically
-      return;
-    }
-    if (currentTrack?.albumId) {
-      handleSetAccentColorOverride(currentTrack.albumId, color);
-    }
-    setAccentColor(color);
-  }, [currentTrack?.albumId, handleSetAccentColorOverride, handleResetAccentColorOverride, setAccentColor]);
-
-  // --- Queue visibility ---
-  const handleShowQueue = useCallback(() => {
-    handlers.onCloseLibraryDrawer();
-    setShowVisualEffects(false);
-    setShowQueue(prev => !prev);
-  }, [setShowQueue, handlers, setShowVisualEffects]);
-  const handleCloseQueue = useCallback(() => setShowQueue(false), [setShowQueue]);
-
-  const handleOpenQueueFromToast = useCallback(() => {
-    handlers.onCloseLibraryDrawer();
-    setShowVisualEffects(false);
-    setShowQueue(true);
-  }, [handlers, setShowVisualEffects, setShowQueue]);
-
-  const handleAddToQueueWithToast = useCallback(
-    async (playlistId: string, playlistName?: string, provider?: ProviderId) => {
-      if (!handlers.onAddToQueue) return null;
-      const result = await handlers.onAddToQueue(playlistId, playlistName, provider);
-      if (!result || result.added <= 0) return null;
-      const name = result.collectionName?.trim();
-      const title = name && name.length > 0 ? `"${name}"` : 'this collection';
-      setToast({
-        message: `Added ${result.added} ${result.added === 1 ? 'track' : 'tracks'} from ${title} to your`,
-        actionLabel: 'queue',
-        onAction: () => {
-          handleOpenQueueFromToast();
-          setToast(null);
-        },
-      });
-      return result;
-    },
-    [handlers, handleOpenQueueFromToast],
-  );
-
-  // --- App settings handlers ---
-  const handleClearCache = useCallback(async (options: ClearCacheOptions) => {
-    await clearCacheWithOptions({ clearLikes: options.clearLikes });
-    if (options.clearPins) {
-      await clearAllPins();
-    }
-    if (options.clearAccentColors) {
-      localStorage.removeItem(STORAGE_KEYS.ACCENT_COLOR_OVERRIDES);
-      localStorage.removeItem(STORAGE_KEYS.CUSTOM_ACCENT_COLORS);
-    }
-    if (options.clearPins || options.clearAccentColors) {
-      const { clearPreferencesSyncTimestamp, getPreferencesSync } =
-        await import('@/providers/dropbox/dropboxPreferencesSync');
-      clearPreferencesSyncTimestamp();
-      getPreferencesSync()?.initialSync();
-    }
-  }, []);
-
-  const handleProfilerToggle = useCallback(() => {
-    if (!profilerEnabled) {
-      vizDebugCtx?.setIsDebugMode(false);
-    }
-    profilerToggle();
-  }, [profilerEnabled, profilerToggle, vizDebugCtx]);
-
-  const handleVisualizerDebugToggle = useCallback(() => {
-    if (!visualizerDebugEnabled && profilerEnabled) {
-      profilerToggle();
-    }
-    vizDebugCtx?.setIsDebugMode(prev => !prev);
-  }, [visualizerDebugEnabled, profilerEnabled, profilerToggle, vizDebugCtx]);
-
-  // --- VFX menu visibility ---
-  const handleShowVisualEffects = useCallback(() => {
-    setShowQueue(false);
-    handlers.onCloseLibraryDrawer();
-    setShowVisualEffects(true);
-  }, [setShowQueue, handlers, setShowVisualEffects]);
-  const handleCloseVisualEffects = useCallback(() => setShowVisualEffects(false), [setShowVisualEffects]);
-  const handleToggleVisualEffectsMenu = useCallback(() => {
-    setShowQueue(false);
-    handlers.onCloseLibraryDrawer();
-    setShowVisualEffects(prev => !prev);
-  }, [setShowVisualEffects, setShowQueue, handlers]);
-
-  // --- Glow toggle (re-enables VFX + restores saved state) ---
-  const handleGlowToggle = useCallback(() => {
-    if (visualEffectsEnabled) {
-      setVisualEffectsEnabled(false);
-    } else {
-      setVisualEffectsEnabled(true);
-      restoreGlowSettings();
-    }
-  }, [visualEffectsEnabled, setVisualEffectsEnabled, restoreGlowSettings]);
-
-  // --- Background visualizer controls ---
-  const handleBackgroundVisualizerToggle = useCallback(() => {
-    setBackgroundVisualizerEnabled(prev => !prev);
-  }, [setBackgroundVisualizerEnabled]);
-
-  const handleBackgroundVisualizerStyleChange = useCallback((style: 'fireflies' | 'comet') => {
-    setBackgroundVisualizerStyle(style);
-  }, [setBackgroundVisualizerStyle]);
-
-  const handleTranslucenceToggle = useCallback(() => {
-    setTranslucenceEnabled(prev => !prev);
-  }, [setTranslucenceEnabled]);
-
-  const handleBackgroundVisualizerIntensityChange = useCallback((intensity: number) => {
-    setBackgroundVisualizerIntensity(Math.max(0, Math.min(100, intensity)));
-  }, [setBackgroundVisualizerIntensity]);
-
-  // --- Library drawer ---
-  const handleOpenLibraryDrawer = useCallback(() => {
-    setShowQueue(false);
-    setShowVisualEffects(false);
-    handlers.onOpenLibraryDrawer();
-  }, [handlers, setShowQueue, setShowVisualEffects]);
-
-  const handleArtistBrowse = useCallback((artistName: string) => {
-    setLibrarySearchQuery(artistName);
-    setLibraryViewMode('albums');
-    handleOpenLibraryDrawer();
-  }, [handleOpenLibraryDrawer]);
-
-  const handleAlbumPlay = useCallback((albumId: string, albumName: string) => {
-    handlers.onAlbumPlay(albumId, albumName);
-  }, [handlers]);
-
-  const handleCloseLibraryDrawer = useCallback(() => {
-    handlers.onCloseLibraryDrawer();
-    setTimeout(() => {
-      setLibrarySearchQuery(undefined);
-      setLibraryViewMode(undefined);
-    }, 350);
-  }, [handlers]);
-
-  // --- Responsive sizing ---
-  const { dimensions, useFluidSizing, padding, transitionDuration, transitionEasing, isMobile, isTablet, hasPointerInput, isTouchDevice } = usePlayerSizingContext();
-
-  // Measure controls card height so album art fills remaining space exactly.
-  const controlsRef = useRef<HTMLDivElement>(null);
-  const stableControlsHeightRef = useRef<number>(220);
-  useEffect(() => {
-    const el = controlsRef.current;
-    if (!el) return;
-    const update = () => {
-      const h = Math.ceil(el.getBoundingClientRect().height);
-      if (h > 50) {
-        stableControlsHeightRef.current = h;
-        document.documentElement.style.setProperty('--player-controls-height', `${h}px`);
-      }
-    };
-    const observer = new ResizeObserver(update);
-    observer.observe(el);
-    update();
-    return () => observer.disconnect();
-  }, []);
-
-  // Pre-set --player-controls-height synchronously before exiting zen so the
-  // transition starts toward the correct target immediately.
-  const handleZenModeToggle = useCallback(() => {
-    if (zenModeEnabled) {
-      document.documentElement.style.setProperty(
-        '--player-controls-height',
-        `${stableControlsHeightRef.current}px`
-      );
-    }
-    setZenModeEnabled(prev => !prev);
-  }, [zenModeEnabled, setZenModeEnabled]);
-
-  // Click outside album art exits zen mode
-  const handleZenExitClick = useCallback((e: React.MouseEvent<HTMLDivElement>) => {
-    if (zenModeEnabled && e.currentTarget.contains(e.target as Node)) {
-      handleZenModeToggle();
-    }
-  }, [zenModeEnabled, handleZenModeToggle]);
-
-  // --- Swipe gestures ---
-  const { offsetX, isSwiping, isAnimating, gestureHandlers } = useSwipeGesture({
-    onSwipeLeft: handlers.onNext,
-    onSwipeRight: handlers.onPrevious,
-  }, { enabled: isTouchDevice });
-
-  const handleSwipeUp = useCallback(() => {
-    if (showQueue) {
-      handleCloseQueue();
-    } else {
-      handleShowQueue();
-    }
-  }, [showQueue, handleShowQueue, handleCloseQueue]);
-
-  const handleSwipeDown = useCallback(() => {
-    if (showLibraryDrawer) {
-      handlers.onCloseLibraryDrawer();
-    } else {
-      handleOpenLibraryDrawer();
-    }
-  }, [showLibraryDrawer, handlers, handleOpenLibraryDrawer]);
-
-  const { ref: drawerSwipeRef } = useVerticalSwipeGesture({
-    onSwipeUp: handleSwipeUp,
-    onSwipeDown: handleSwipeDown,
-    threshold: 80,
-    enabled: isTouchDevice,
-  });
-
-  // --- Keyboard shortcut helpers ---
-  const handleEscapeClose = useCallback(() => {
-    handleCloseLibraryDrawer();
-    handleCloseVisualEffects();
-    if (showHelp) closeHelp();
-    setIsFlipped(false);
-  }, [handleCloseLibraryDrawer, handleCloseVisualEffects, showHelp, closeHelp]);
-
-
-  const handlePlayPause = useCallback(() => {
-    if (isPlaying) {
-      handlers.onPause();
-    } else {
-      handlers.onPlay();
-    }
-  }, [handlers, isPlaying]);
-
-  const handleArrowUp = useCallback(() => {
-    if (showLibraryDrawer) {
-      handlers.onCloseLibraryDrawer();
-      handleShowQueue();
-    } else if (showQueue) {
-      handleCloseQueue();
-    } else {
-      handleShowQueue();
-    }
-  }, [handlers, showLibraryDrawer, showQueue, handleShowQueue, handleCloseQueue]);
-
-  const handleArrowDown = useCallback(() => {
-    if (showQueue) {
-      handleCloseQueue();
-      handleOpenLibraryDrawer();
-    } else if (showLibraryDrawer) {
-      handlers.onCloseLibraryDrawer();
-    } else {
-      handleOpenLibraryDrawer();
-    }
-  }, [handlers, showQueue, showLibraryDrawer, handleCloseQueue, handleOpenLibraryDrawer]);
-
-  const handleVolumeUp = useCallback(() => {
-    setVolumeLevel(Math.min(100, (volume ?? 50) + 5));
-  }, [volume, setVolumeLevel]);
-
-  const handleVolumeDown = useCallback(() => {
-    setVolumeLevel(Math.max(0, (volume ?? 50) - 5));
-  }, [volume, setVolumeLevel]);
-
-  useKeyboardShortcuts({
-    onPlayPause: handlePlayPause,
-    onNext: handlers.onNext,
-    onPrevious: handlers.onPrevious,
-    onCloseQueue: handleCloseQueue,
-    onToggleVisualEffectsMenu: handleToggleVisualEffectsMenu,
-    onCloseVisualEffects: handleEscapeClose,
-    onToggleBackgroundVisualizer: handleBackgroundVisualizerToggle,
-    onToggleGlow: handleGlowToggle,
-    onToggleTranslucence: handleTranslucenceToggle,
-    onMute: handleMuteToggle,
-    onVolumeUp: handleVolumeUp,
-    onVolumeDown: handleVolumeDown,
-    onToggleLike: handleLikeToggle,
-    onToggleShuffle: handleShuffleToggle,
-    onToggleHelp: toggleHelp,
-    onShowQueue: handleArrowUp,
-    onOpenLibraryDrawer: handleArrowDown,
-    onToggleZenMode: handleZenModeToggle,
-  }, { prefersPointerInput: hasPointerInput });
-
-  return (
-    <ContentWrapper
-      width={dimensions.width}
-      padding={padding}
-      useFluidSizing={useFluidSizing}
-      transitionDuration={transitionDuration}
-      transitionEasing={transitionEasing}
-      $zenMode={zenModeEnabled}
-      onClick={handleZenExitClick}
-    >
-      <PlayerContainer>
-        <PlayerStack $zenMode={zenModeEnabled}>
-          {/* Album Art Zone */}
-          <CardContent style={{
-            position: 'relative',
-            zIndex: 2,
-            minHeight: 0,
-            alignItems: 'center',
-            paddingTop: zenModeEnabled ? '0' : (isMobile ? '0.25rem' : '0.5rem')
-          }}>
-            <div ref={flipContainerRef} style={{ width: '100%' }}>
-              <ClickableAlbumArtContainer
-                ref={(el) => {
-                  albumArtContainerRef.current = el;
-                  if (isTouchDevice && drawerSwipeRef && typeof drawerSwipeRef !== 'function') {
-                    (drawerSwipeRef as React.MutableRefObject<HTMLDivElement | null>).current = el;
-                  } else if (isTouchDevice && typeof drawerSwipeRef === 'function') {
-                    (drawerSwipeRef as (instance: HTMLDivElement | null) => void)(el);
-                  }
-                }}
-                $swipeEnabled={isTouchDevice}
-                $bothGestures={isTouchDevice}
-                {...(isTouchDevice ? gestureHandlers : {})}
-                onClick={(e) => {
-                  e.stopPropagation();
-                  if (!isSwiping && !isAnimating && !isFlipped) toggleFlip();
-                }}
-                style={{
-                  transform: `translateX(${offsetX}px)`,
-                  transition: isAnimating ? 'transform 300ms cubic-bezier(0.4, 0, 0.2, 1)' : 'none',
-                  willChange: isSwiping ? 'transform' : undefined,
-                }}
-              >
-                <FlipInner $isFlipped={isFlipped}>
-                  <ProfiledComponent id="AlbumArt">
-                    <AlbumArt
-                      currentTrack={currentTrack}
-                      accentColor={accentColor}
-                      glowIntensity={visualEffectsEnabled ? effectiveGlow.intensity : 0}
-                      glowRate={effectiveGlow.rate}
-                      glowEnabled={visualEffectsEnabled}
-                      translucenceEnabled={translucenceEnabled}
-                      translucenceOpacity={translucenceOpacity}
-                      zenMode={zenModeEnabled}
-                      onRetryAlbumArt={handleRetryAlbumArt}
-                    />
-                  </ProfiledComponent>
-                  <AlbumArtQuickSwapBack
-                    currentTrack={currentTrack}
-                    accentColor={accentColor}
-                    onAccentColorChange={handleAccentColorChange}
-                    customAccentColorOverrides={customAccentColors}
-                    onCustomAccentColor={handleCustomAccentColor}
-                    glowEnabled={visualEffectsEnabled}
-                    onGlowToggle={handleGlowToggle}
-                    glowIntensity={effectiveGlow.intensity}
-                    onGlowIntensityChange={handleGlowIntensityChange}
-                    glowRate={effectiveGlow.rate}
-                    onGlowRateChange={handleGlowRateChange}
-                    backgroundVisualizerEnabled={backgroundVisualizerEnabled}
-                    onBackgroundVisualizerToggle={handleBackgroundVisualizerToggle}
-                    backgroundVisualizerStyle={backgroundVisualizerStyle as 'fireflies' | 'comet'}
-                    onBackgroundVisualizerStyleChange={handleBackgroundVisualizerStyleChange}
-                    backgroundVisualizerIntensity={backgroundVisualizerIntensity}
-                    onBackgroundVisualizerIntensityChange={handleBackgroundVisualizerIntensityChange}
-                    translucenceEnabled={translucenceEnabled}
-                    onTranslucenceToggle={handleTranslucenceToggle}
-                    isMobile={isMobile}
-                    isTablet={isTablet}
-                    onClose={() => setIsFlipped(false)}
-                    onRetryAlbumArt={handleRetryAlbumArt}
-                  />
-                </FlipInner>
-              </ClickableAlbumArtContainer>
-            </div>
-          </CardContent>
-          <ZenTrackInfo $zenMode={zenModeEnabled}>
-            <ZenTrackName $isMobile={isMobile} $isTablet={isTablet}>
-              {currentTrack?.name}
-            </ZenTrackName>
-            {currentTrack?.artists && (
-              <ZenTrackArtist>{currentTrack.artists}</ZenTrackArtist>
-            )}
-          </ZenTrackInfo>
-          <ZenControlsWrapper $zenMode={zenModeEnabled}>
-            <ZenControlsInner ref={controlsRef}>
-              <LoadingCard
-                backgroundImage={currentTrack?.image}
-                accentColor={accentColor}
-                glowEnabled={visualEffectsEnabled}
-                glowIntensity={effectiveGlow.intensity}
-                glowRate={effectiveGlow.rate}
-              >
-                <CardContent style={{
-                  position: 'relative',
-                  zIndex: 2,
-                  flex: '0 0 auto',
-                  minHeight: 'auto',
-                  display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'center'
-                }}>
-                  <Suspense fallback={<ControlsLoadingFallback />}>
-                    <ProfiledComponent id="SpotifyPlayerControls">
-                      <SpotifyPlayerControls
-                        currentTrack={currentTrack}
-                        trackCount={tracks.length}
-                        isLiked={isLiked}
-                        isLikePending={isLikePending}
-                        onToggleLike={handleLikeToggle}
-                        onPlay={handlers.onPlay}
-                        onPause={handlers.onPause}
-                        onNext={handlers.onNext}
-                        onPrevious={handlers.onPrevious}
-                        onArtistBrowse={handleArtistBrowse}
-                        onAlbumPlay={handleAlbumPlay}
-                        currentTrackProvider={currentTrackProvider}
-                      />
-                    </ProfiledComponent>
-                  </Suspense>
-                </CardContent>
-              </LoadingCard>
-            </ZenControlsInner>
-          </ZenControlsWrapper>
-        </PlayerStack>
-      </PlayerContainer>
-      <ProfiledComponent id="BottomBar">
-      <BottomBar
-        zenModeEnabled={zenModeEnabled}
-        isMuted={isMuted}
-        volume={volume}
-        onMuteToggle={handleMuteToggle}
-        onVolumeChange={setVolumeLevel}
-        onShowVisualEffects={handleShowVisualEffects}
-        onBackToLibrary={handleOpenLibraryDrawer}
-        onShowQueue={handleShowQueue}
-        onZenModeToggle={handleZenModeToggle}
-        shuffleEnabled={shuffleEnabled}
-        onShuffleToggle={handleShuffleToggle}
-        onStartRadio={isRadioAvailable ? handlers.onStartRadio : undefined}
-        radioGenerating={radioState?.isGenerating}
-      />
-      </ProfiledComponent>
-      <Suspense fallback={<VisualEffectsLoadingFallback />}>
-        <VisualEffectsMenu
-          isOpen={showVisualEffects}
-          onClose={handleCloseVisualEffects}
-          onClearCache={handleClearCache}
-          profilerEnabled={profilerEnabled}
-          onProfilerToggle={handleProfilerToggle}
-          visualizerDebugEnabled={visualizerDebugEnabled}
-          onVisualizerDebugToggle={handleVisualizerDebugToggle}
-        />
-      </Suspense>
-      <Suspense fallback={<QueueLoadingFallback />}>
-        {isMobile ? (
-          <ProfiledComponent id="QueueBottomSheet">
-            <QueueBottomSheet
-              isOpen={showQueue}
-              onClose={handleCloseQueue}
-              tracks={tracks}
-              currentTrackIndex={currentTrackIndex}
-              onTrackSelect={handlers.onTrackSelect}
-              onRemoveTrack={handlers.onRemoveFromQueue}
-              onReorderTracks={handlers.onReorderQueue}
-              showProviderIcons={showProviderIcons}
-              radioActive={radioActive}
-              radioSeedDescription={radioState?.seedDescription}
-              onSaveQueue={handleOpenSaveQueue}
-              canSaveQueue={canSaveQueue}
-            />
-          </ProfiledComponent>
-        ) : (
-          <ProfiledComponent id="QueueDrawer">
-            <QueueDrawer
-              isOpen={showQueue}
-              onClose={handleCloseQueue}
-              tracks={tracks}
-              currentTrackIndex={currentTrackIndex}
-              onTrackSelect={handlers.onTrackSelect}
-              onRemoveTrack={handlers.onRemoveFromQueue}
-              onReorderTracks={handlers.onReorderQueue}
-              showProviderIcons={showProviderIcons}
-              radioActive={radioActive}
-              radioSeedDescription={radioState?.seedDescription}
-              onSaveQueue={handleOpenSaveQueue}
-              canSaveQueue={canSaveQueue}
-            />
-          </ProfiledComponent>
-        )}
-      </Suspense>
-      <Suspense fallback={null}>
-        <KeyboardShortcutsHelp isOpen={showHelp} onClose={closeHelp} />
-      </Suspense>
-      <ProfiledComponent id="LibraryDrawer">
-        <LibraryDrawer
-          isOpen={showLibraryDrawer}
-          onClose={handleCloseLibraryDrawer}
-          onPlaylistSelect={handlers.onPlaylistSelect}
-          onAddToQueue={handleAddToQueueWithToast}
-          initialSearchQuery={librarySearchQuery}
-          initialViewMode={libraryViewMode}
-        />
-      </ProfiledComponent>
-      {showSaveQueueDialog && (
-        <Suspense fallback={null}>
-          <SaveQueueDialog
-            onSave={handleSaveQueue}
-            onClose={handleCloseSaveQueue}
-            availableProviders={saveProviders}
-            trackProviders={trackProviders}
-            defaultName={radioState?.isActive && radioState.seedDescription ? radioState.seedDescription.replace(/^Radio based on /i, '').replace(/\s+by\s+.+$/i, '') + ' Radio' : undefined}
-          />
-        </Suspense>
-      )}
-      {toast && (
-        <Toast
-          message={toast.message}
-          onDismiss={handleDismissToast}
-          actionLabel={toast.actionLabel}
-          onAction={toast.onAction}
-        />
-      )}
-      {radioProgress && onDismissRadioProgress && (
-        <RadioProgressToast
-          phase={radioProgress.phase}
-          trackCount={radioProgress.trackCount}
-          onDismiss={onDismissRadioProgress}
-          onViewQueue={() => {
-            handleOpenQueueFromToast();
-            onDismissRadioProgress();
-          }}
-        />
-      )}
-    </ContentWrapper>
-  );
-});
-
-PlayerContent.displayName = 'PlayerContent';
-
-export default PlayerContent;
