+++ src/components/PlayerContent/AlbumArtSection.tsx
@@ -1,0 +1,258 @@
+import React, { useState, useCallback, useRef, useEffect } from 'react';
+import { CardContent } from '@/components/styled';
+import AlbumArt from '@/components/AlbumArt';
+import AlbumArtQuickSwapBack from '@/components/AlbumArtQuickSwapBack';
+import { ProfiledComponent } from '@/components/ProfiledComponent';
+import { useColorContext } from '@/contexts/ColorContext';
+import { useVisualEffectsContext } from '@/contexts/VisualEffectsContext';
+import { useVisualEffectsState } from '@/hooks/useVisualEffectsState';
+import type { MediaTrack, ProviderId } from '@/types/domain';
+import { FlipInner, ZenTrackInfo, ZenTrackName, ZenTrackArtist } from './styled';
+import { GestureLayer } from './GestureLayer';
+
+interface AlbumArtBounds {
+  left: number;
+  top: number;
+  width: number;
+  height: number;
+}
+
+interface AlbumArtSectionProps {
+  currentTrack: MediaTrack | null;
+  currentTrackProvider?: ProviderId;
+  zenModeEnabled: boolean;
+  isMobile: boolean;
+  isTablet: boolean;
+  isTouchDevice: boolean;
+  onSwipeLeft: () => void;
+  onSwipeRight: () => void;
+  onSwipeUp: () => void;
+  onSwipeDown: () => void;
+  onAlbumArtBoundsChange?: (bounds: AlbumArtBounds | null) => void;
+}
+
+export const AlbumArtSection: React.FC<AlbumArtSectionProps> = React.memo(({
+  currentTrack,
+  currentTrackProvider,
+  zenModeEnabled,
+  isMobile,
+  isTablet,
+  isTouchDevice,
+  onSwipeLeft,
+  onSwipeRight,
+  onSwipeUp,
+  onSwipeDown,
+  onAlbumArtBoundsChange,
+}) => {
+  const {
+    accentColor,
+    customAccentColors,
+    setAccentColor,
+    handleSetAccentColorOverride,
+    handleRemoveAccentColorOverride,
+    handleResetAccentColorOverride,
+    handleSetCustomAccentColor,
+    handleRemoveCustomAccentColor,
+  } = useColorContext();
+
+  const {
+    visualEffectsEnabled,
+    translucenceEnabled,
+    translucenceOpacity,
+    backgroundVisualizerEnabled,
+    setBackgroundVisualizerEnabled,
+    backgroundVisualizerStyle,
+    setBackgroundVisualizerStyle,
+    backgroundVisualizerIntensity,
+    setBackgroundVisualizerIntensity,
+    setTranslucenceEnabled,
+    setVisualEffectsEnabled,
+  } = useVisualEffectsContext();
+
+  const {
+    effectiveGlow,
+    handleGlowIntensityChange,
+    handleGlowRateChange,
+    restoreGlowSettings,
+  } = useVisualEffectsState();
+
+  const [isFlipped, setIsFlipped] = useState(false);
+  const flipContainerRef = useRef<HTMLDivElement>(null);
+  const albumArtContainerRef = useRef<HTMLDivElement | null>(null);
+
+  const toggleFlip = useCallback(() => setIsFlipped(f => !f), []);
+
+  const handleRetryAlbumArt = useCallback(async () => {
+    const providerId = currentTrackProvider ?? currentTrack?.provider;
+    if (!providerId) return;
+    const { providerRegistry } = await import('@/providers/registry');
+    const descriptor = providerRegistry.get(providerId);
+    descriptor?.playback.refreshCurrentTrackArt?.();
+  }, [currentTrackProvider, currentTrack?.provider]);
+
+  useEffect(() => {
+    if (!onAlbumArtBoundsChange) return;
+    const el = albumArtContainerRef.current;
+    if (!el) {
+      onAlbumArtBoundsChange(null);
+      return;
+    }
+    const report = () => {
+      const r = el.getBoundingClientRect();
+      if (r.width > 0 && r.height > 0) {
+        onAlbumArtBoundsChange({ left: r.left, top: r.top, width: r.width, height: r.height });
+      } else {
+        onAlbumArtBoundsChange(null);
+      }
+    };
+    report();
+    const ro = new ResizeObserver(report);
+    ro.observe(el);
+    return () => {
+      ro.disconnect();
+      onAlbumArtBoundsChange(null);
+    };
+  }, [onAlbumArtBoundsChange]);
+
+  useEffect(() => {
+    setIsFlipped(false);
+  }, [currentTrack?.id]);
+
+  useEffect(() => {
+    if (!isFlipped) return;
+    const handleClickOutside = (e: MouseEvent) => {
+      const target = e.target as Element;
+      if (target.closest?.('[data-eyedropper-overlay]')) return;
+      const insideAlbumArt = flipContainerRef.current?.contains(target);
+      if (!insideAlbumArt) {
+        setIsFlipped(false);
+      }
+    };
+    document.addEventListener('mousedown', handleClickOutside);
+    return () => document.removeEventListener('mousedown', handleClickOutside);
+  }, [isFlipped]);
+
+  const handleCustomAccentColor = useCallback((color: string) => {
+    if (currentTrack?.albumId) {
+      if (color === '') {
+        handleRemoveAccentColorOverride(currentTrack.albumId);
+        handleRemoveCustomAccentColor(currentTrack.albumId);
+      } else {
+        handleSetAccentColorOverride(currentTrack.albumId, color);
+        handleSetCustomAccentColor(currentTrack.albumId, color);
+        setAccentColor(color);
+      }
+    }
+  }, [currentTrack?.albumId, handleSetAccentColorOverride, handleRemoveAccentColorOverride, handleSetCustomAccentColor, handleRemoveCustomAccentColor, setAccentColor]);
+
+  const handleAccentColorChange = useCallback((color: string) => {
+    if (color === 'RESET_TO_DEFAULT' && currentTrack?.albumId) {
+      handleResetAccentColorOverride(currentTrack.albumId);
+      return;
+    }
+    if (currentTrack?.albumId) {
+      handleSetAccentColorOverride(currentTrack.albumId, color);
+    }
+    setAccentColor(color);
+  }, [currentTrack?.albumId, handleSetAccentColorOverride, handleResetAccentColorOverride, setAccentColor]);
+
+  const handleGlowToggle = useCallback(() => {
+    if (visualEffectsEnabled) {
+      setVisualEffectsEnabled(false);
+    } else {
+      setVisualEffectsEnabled(true);
+      restoreGlowSettings();
+    }
+  }, [visualEffectsEnabled, setVisualEffectsEnabled, restoreGlowSettings]);
+
+  const handleBackgroundVisualizerToggle = useCallback(() => {
+    setBackgroundVisualizerEnabled(prev => !prev);
+  }, [setBackgroundVisualizerEnabled]);
+
+  const handleBackgroundVisualizerStyleChange = useCallback((style: 'fireflies' | 'comet') => {
+    setBackgroundVisualizerStyle(style);
+  }, [setBackgroundVisualizerStyle]);
+
+  const handleTranslucenceToggle = useCallback(() => {
+    setTranslucenceEnabled(prev => !prev);
+  }, [setTranslucenceEnabled]);
+
+  const handleBackgroundVisualizerIntensityChange = useCallback((intensity: number) => {
+    setBackgroundVisualizerIntensity(Math.max(0, Math.min(100, intensity)));
+  }, [setBackgroundVisualizerIntensity]);
+
+  return (
+    <>
+      <CardContent style={{
+        position: 'relative',
+        zIndex: 2,
+        minHeight: 0,
+        alignItems: 'center',
+        paddingTop: zenModeEnabled ? '0' : (isMobile ? '0.25rem' : '0.5rem')
+      }}>
+        <div ref={flipContainerRef} style={{ width: '100%' }}>
+          <GestureLayer
+            onSwipeLeft={onSwipeLeft}
+            onSwipeRight={onSwipeRight}
+            onSwipeUp={onSwipeUp}
+            onSwipeDown={onSwipeDown}
+            isTouchDevice={isTouchDevice}
+            onClick={toggleFlip}
+            albumArtContainerRef={albumArtContainerRef}
+          >
+            <FlipInner $isFlipped={isFlipped}>
+              <ProfiledComponent id="AlbumArt">
+                <AlbumArt
+                  currentTrack={currentTrack}
+                  accentColor={accentColor}
+                  glowIntensity={visualEffectsEnabled ? effectiveGlow.intensity : 0}
+                  glowRate={effectiveGlow.rate}
+                  glowEnabled={visualEffectsEnabled}
+                  translucenceEnabled={translucenceEnabled}
+                  translucenceOpacity={translucenceOpacity}
+                  zenMode={zenModeEnabled}
+                  onRetryAlbumArt={handleRetryAlbumArt}
+                />
+              </ProfiledComponent>
+              <AlbumArtQuickSwapBack
+                currentTrack={currentTrack}
+                accentColor={accentColor}
+                onAccentColorChange={handleAccentColorChange}
+                customAccentColorOverrides={customAccentColors}
+                onCustomAccentColor={handleCustomAccentColor}
+                glowEnabled={visualEffectsEnabled}
+                onGlowToggle={handleGlowToggle}
+                glowIntensity={effectiveGlow.intensity}
+                onGlowIntensityChange={handleGlowIntensityChange}
+                glowRate={effectiveGlow.rate}
+                onGlowRateChange={handleGlowRateChange}
+                backgroundVisualizerEnabled={backgroundVisualizerEnabled}
+                onBackgroundVisualizerToggle={handleBackgroundVisualizerToggle}
+                backgroundVisualizerStyle={backgroundVisualizerStyle as 'fireflies' | 'comet'}
+                onBackgroundVisualizerStyleChange={handleBackgroundVisualizerStyleChange}
+                backgroundVisualizerIntensity={backgroundVisualizerIntensity}
+                onBackgroundVisualizerIntensityChange={handleBackgroundVisualizerIntensityChange}
+                translucenceEnabled={translucenceEnabled}
+                onTranslucenceToggle={handleTranslucenceToggle}
+                isMobile={isMobile}
+                isTablet={isTablet}
+                onClose={() => setIsFlipped(false)}
+                onRetryAlbumArt={handleRetryAlbumArt}
+              />
+            </FlipInner>
+          </GestureLayer>
+        </div>
+      </CardContent>
+      <ZenTrackInfo $zenMode={zenModeEnabled}>
+        <ZenTrackName $isMobile={isMobile} $isTablet={isTablet}>
+          {currentTrack?.name}
+        </ZenTrackName>
+        {currentTrack?.artists && (
+          <ZenTrackArtist>{currentTrack.artists}</ZenTrackArtist>
+        )}
+      </ZenTrackInfo>
+    </>
+  );
+});
+
+AlbumArtSection.displayName = 'AlbumArtSection';
